1.参数规格查询

+++ src/item/desc/TbItemDesc.go
@@ -7,7 +7,7 @@ package desc
  */
 
 type TbItemDesc struct {
-	ItemId   int
+	ItemId   int64
 	ItemDesc string
 	Created  string
 	Update   string
